Rename modules file.
